+++ src/app/components/shared/navbar/navbar.component.spec.ts
@@ -1,6 +1,7 @@
 import {async, ComponentFixture, TestBed} from '@angular/core/testing';
 
 import {NavbarComponent} from './navbar.component';
+import {AppModule} from "../../../app.module";
 
 describe('NavbarComponent', () => {
   let component: NavbarComponent;
@@ -8,6 +9,7 @@ describe('NavbarComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
+      imports: [AppModule],
       declarations: [NavbarComponent]
     })
       .compileComponents();